Adding an unit test target for gnews api calls
