+++ TrendingNewsML/Podfile
@@ -20,6 +20,15 @@ target 'TrendingNewsMLUnitTests' do
 
 end
 
+target 'TrendingNewsMLGNewsUnitTests' do
+  # Comment the next line if you don't want to use dynamic frameworks
+  use_frameworks!
+
+  # Pods for TrendingNewsML
+  pod 'ReadabilityKit'
+
+end
+
 target 'TrendingNewsMLUITests' do
   # Comment the next line if you don't want to use dynamic frameworks
   use_frameworks!
