+++ TrendingNewsML/TrendingNewsMLUnitTests/TrendingNewsMLUnitTests.swift
@@ -7,7 +7,6 @@
 //
 
 import XCTest
-@testable import TrendingNewsML
 
 class TrendingNewsMLUnitTests: XCTestCase {
 
@@ -28,22 +27,6 @@ class TrendingNewsMLUnitTests: XCTestCase {
         // This is an example of a performance test case.
         measure {
             // Put the code you want to measure the time of here.
-        }
-    }
-    
-    func testArticlesRequest() {
-        let promise = expectation(description: "Get list of articles of category")
-        let timeout: TimeInterval = 5 // 5 seconds
-        
-        GNews.getArticles(of: .business) { articles in
-            XCTAssertGreaterThan(articles.count, 0)
-            promise.fulfill()
-        }
-        
-        waitForExpectations(timeout: timeout) { error in
-            if let error = error {
-                XCTFail("waitForExpectationsWithTimeout errored: \(error)")
-            }
         }
     }
 }
